+++ lib/dictosaurus.dart
@@ -64,3 +64,4 @@ export 'src/dictionary/_index.dart'
 export 'src/dictosaurus/_index.dart'
     show DictoSaurus, DictoSaurusMixin, DictoSaurusBase;
 export 'src/_common/_index.dart' show TermProperty, PartOfSpeech;
+export 'package:gmconsult_dart_core/dart_core.dart' show Language;